refactor : DetailSection 폴더 변경
