+++ components/detailSection/DetailSectionheader.tsx
